+++ crates/rust-client/src/store/web_store/notes/mod.rs
@@ -3,18 +3,15 @@ use alloc::{
     vec::Vec,
 };
 
+use js_sys::{Array, Promise};
 use miden_objects::{notes::Nullifier, Digest};
 use serde_wasm_bindgen::from_value;
+use wasm_bindgen::JsValue;
 use wasm_bindgen_futures::*;
 
 use super::WebStore;
 use crate::store::{
-    note_record::{
-        STATE_COMMITTED, STATE_CONSUMED_AUTHENTICATED_LOCAL, STATE_CONSUMED_EXTERNAL,
-        STATE_CONSUMED_UNAUTHENTICATED_LOCAL, STATE_EXPECTED, STATE_PROCESSING_AUTHENTICATED,
-        STATE_PROCESSING_UNAUTHENTICATED, STATE_UNVERIFIED,
-    },
-    InputNoteRecord, NoteFilter, OutputNoteRecord, StoreError,
+    InputNoteRecord, InputNoteState, NoteFilter, OutputNoteRecord, OutputNoteState, StoreError,
 };
 
 mod js_bindings;
@@ -31,55 +28,7 @@ impl WebStore {
         &self,
         filter: NoteFilter,
     ) -> Result<Vec<InputNoteRecord>, StoreError> {
-        let promise = match &filter {
-            NoteFilter::All
-            | NoteFilter::Consumed
-            | NoteFilter::Committed
-            | NoteFilter::Expected
-            | NoteFilter::Processing
-            | NoteFilter::Unverified => {
-                let states: Vec<u8> = match filter {
-                    NoteFilter::All => vec![],
-                    NoteFilter::Consumed => vec![
-                        STATE_CONSUMED_AUTHENTICATED_LOCAL,
-                        STATE_CONSUMED_UNAUTHENTICATED_LOCAL,
-                        STATE_CONSUMED_EXTERNAL,
-                    ],
-                    NoteFilter::Committed => vec![STATE_COMMITTED],
-                    NoteFilter::Expected => vec![STATE_EXPECTED],
-                    NoteFilter::Processing => {
-                        vec![STATE_PROCESSING_AUTHENTICATED, STATE_PROCESSING_UNAUTHENTICATED]
-                    },
-                    NoteFilter::Unverified => vec![STATE_UNVERIFIED],
-                    _ => unreachable!(), // Safety net, should never be reached
-                };
-
-                // Assuming `js_fetch_notes` is your JavaScript function that handles simple string
-                // filters
-                idxdb_get_input_notes(states)
-            },
-            NoteFilter::Ignored => idxdb_get_input_notes(vec![]),
-            NoteFilter::List(ids) => {
-                let note_ids_as_str: Vec<String> =
-                    ids.iter().map(|id| id.inner().to_string()).collect();
-                idxdb_get_input_notes_from_ids(note_ids_as_str)
-            },
-            NoteFilter::Unique(id) => {
-                let note_id_as_str = id.inner().to_string();
-                let note_ids = vec![note_id_as_str];
-                idxdb_get_input_notes_from_ids(note_ids)
-            },
-            NoteFilter::Nullifiers(nullifiers) => {
-                let nullifiers_as_str = nullifiers
-                    .iter()
-                    .map(|nullifier| nullifier.to_string())
-                    .collect::<Vec<String>>();
-
-                idxdb_get_input_notes_from_nullifiers(nullifiers_as_str)
-            },
-        };
-
-        let js_value = JsFuture::from(promise).await.unwrap();
+        let js_value = JsFuture::from(filter.to_input_notes_promise()).await.unwrap();
         let input_notes_idxdb: Vec<InputNoteIdxdbObject> = from_value(js_value).unwrap();
 
         let native_input_notes: Result<Vec<InputNoteRecord>, StoreError> = input_notes_idxdb
@@ -113,43 +62,7 @@ impl WebStore {
         &self,
         filter: NoteFilter,
     ) -> Result<Vec<OutputNoteRecord>, StoreError> {
-        let promise = match filter {
-            NoteFilter::All
-            | NoteFilter::Consumed
-            | NoteFilter::Committed
-            | NoteFilter::Expected
-            | NoteFilter::Processing => {
-                let filter_as_str = match filter {
-                    NoteFilter::All => "All",
-                    NoteFilter::Consumed => "Consumed",
-                    NoteFilter::Committed => "Committed",
-                    NoteFilter::Expected => "Expected",
-                    NoteFilter::Processing => "Processing",
-                    _ => unreachable!(), // Safety net, should never be reached
-                };
-
-                // Assuming `js_fetch_notes` is your JavaScript function that handles simple string
-                // filters
-
-                idxdb_get_output_notes(filter_as_str.to_string())
-            },
-            NoteFilter::Ignored => idxdb_get_ignored_output_notes(),
-            NoteFilter::List(ref ids) => {
-                let note_ids_as_str: Vec<String> =
-                    ids.iter().map(|id| id.inner().to_string()).collect();
-                idxdb_get_output_notes_from_ids(note_ids_as_str)
-            },
-            NoteFilter::Unique(id) => {
-                let note_id_as_str = id.inner().to_string();
-                let note_ids = vec![note_id_as_str];
-                idxdb_get_output_notes_from_ids(note_ids)
-            },
-            NoteFilter::Nullifiers(_) | NoteFilter::Unverified => {
-                todo!("Is not currently called, will be implemented in the future");
-            },
-        };
-
-        let js_value = JsFuture::from(promise).await.unwrap();
+        let js_value = JsFuture::from(filter.to_output_note_promise()).await.unwrap();
 
         let output_notes_idxdb: Vec<OutputNoteIdxdbObject> = from_value(js_value).unwrap();
 
@@ -193,7 +106,127 @@ impl WebStore {
             .collect::<Result<Vec<Nullifier>, _>>()
     }
 
-    pub(crate) async fn upsert_input_note(&self, note: InputNoteRecord) -> Result<(), StoreError> {
-        upsert_input_note_tx(note).await
+    pub(crate) async fn upsert_input_notes(
+        &self,
+        notes: &[InputNoteRecord],
+    ) -> Result<(), StoreError> {
+        for note in notes {
+            upsert_input_note_tx(note).await?;
+        }
+
+        Ok(())
+    }
+}
+
+impl NoteFilter {
+    fn to_input_notes_promise(&self) -> Promise {
+        match self {
+            NoteFilter::All
+            | NoteFilter::Consumed
+            | NoteFilter::Committed
+            | NoteFilter::Expected
+            | NoteFilter::Processing
+            | NoteFilter::Unspent
+            | NoteFilter::Unverified => {
+                let states: Vec<u8> = match self {
+                    NoteFilter::All => vec![],
+                    NoteFilter::Consumed => vec![
+                        InputNoteState::STATE_CONSUMED_AUTHENTICATED_LOCAL,
+                        InputNoteState::STATE_CONSUMED_UNAUTHENTICATED_LOCAL,
+                        InputNoteState::STATE_CONSUMED_EXTERNAL,
+                    ],
+                    NoteFilter::Committed => vec![InputNoteState::STATE_COMMITTED],
+                    NoteFilter::Expected => vec![InputNoteState::STATE_EXPECTED],
+                    NoteFilter::Processing => {
+                        vec![
+                            InputNoteState::STATE_PROCESSING_AUTHENTICATED,
+                            InputNoteState::STATE_PROCESSING_UNAUTHENTICATED,
+                        ]
+                    },
+                    NoteFilter::Unverified => vec![InputNoteState::STATE_UNVERIFIED],
+                    NoteFilter::Unspent => vec![
+                        InputNoteState::STATE_EXPECTED,
+                        InputNoteState::STATE_COMMITTED,
+                        InputNoteState::STATE_UNVERIFIED,
+                        InputNoteState::STATE_PROCESSING_AUTHENTICATED,
+                        InputNoteState::STATE_PROCESSING_UNAUTHENTICATED,
+                    ],
+                    _ => unreachable!(), // Safety net, should never be reached
+                };
+
+                // Assuming `js_fetch_notes` is your JavaScript function that handles simple string
+                // filters
+                idxdb_get_input_notes(states)
+            },
+            NoteFilter::List(ids) => {
+                let note_ids_as_str: Vec<String> =
+                    ids.iter().map(|id| id.inner().to_string()).collect();
+                idxdb_get_input_notes_from_ids(note_ids_as_str)
+            },
+            NoteFilter::Unique(id) => {
+                let note_id_as_str = id.inner().to_string();
+                let note_ids = vec![note_id_as_str];
+                idxdb_get_input_notes_from_ids(note_ids)
+            },
+            NoteFilter::Nullifiers(nullifiers) => {
+                let nullifiers_as_str = nullifiers
+                    .iter()
+                    .map(|nullifier| nullifier.to_string())
+                    .collect::<Vec<String>>();
+
+                idxdb_get_input_notes_from_nullifiers(nullifiers_as_str)
+            },
+        }
+    }
+
+    fn to_output_note_promise(&self) -> Promise {
+        match self {
+            NoteFilter::All
+            | NoteFilter::Consumed
+            | NoteFilter::Committed
+            | NoteFilter::Expected
+            | NoteFilter::Unspent => {
+                let states = match self {
+                    NoteFilter::All => vec![],
+                    NoteFilter::Consumed => vec![OutputNoteState::STATE_CONSUMED],
+                    NoteFilter::Committed => vec![
+                        OutputNoteState::STATE_COMMITTED_FULL,
+                        OutputNoteState::STATE_COMMITTED_PARTIAL,
+                    ],
+                    NoteFilter::Expected => vec![
+                        OutputNoteState::STATE_EXPECTED_FULL,
+                        OutputNoteState::STATE_EXPECTED_PARTIAL,
+                    ],
+                    NoteFilter::Unspent => vec![
+                        OutputNoteState::STATE_EXPECTED_FULL,
+                        OutputNoteState::STATE_COMMITTED_FULL,
+                    ],
+                    _ => unreachable!(), // Safety net, should never be reached
+                };
+
+                idxdb_get_output_notes(states)
+            },
+            NoteFilter::Processing | NoteFilter::Unverified => {
+                Promise::resolve(&JsValue::from(Array::new()))
+            },
+            NoteFilter::List(ref ids) => {
+                let note_ids_as_str: Vec<String> =
+                    ids.iter().map(|id| id.inner().to_string()).collect();
+                idxdb_get_output_notes_from_ids(note_ids_as_str)
+            },
+            NoteFilter::Unique(id) => {
+                let note_id_as_str = id.inner().to_string();
+                let note_ids = vec![note_id_as_str];
+                idxdb_get_output_notes_from_ids(note_ids)
+            },
+            NoteFilter::Nullifiers(nullifiers) => {
+                let nullifiers_as_str = nullifiers
+                    .iter()
+                    .map(|nullifier| nullifier.to_string())
+                    .collect::<Vec<String>>();
+
+                idxdb_get_output_notes_from_nullifiers(nullifiers_as_str)
+            },
+        }
     }
 }