+++ crates/rust-client/src/store/web_store/sync/mod.rs
@@ -5,18 +5,23 @@ use alloc::{
 
 use miden_objects::{
     accounts::AccountId,
-    notes::{NoteId, NoteInclusionProof, NoteTag},
+    notes::{NoteId, NoteTag, Nullifier},
 };
 use miden_tx::utils::{Deserializable, Serializable};
 use serde_wasm_bindgen::from_value;
 use wasm_bindgen_futures::*;
 
 use super::{
-    chain_data::utils::serialize_chain_mmr_node, notes::utils::upsert_input_note_tx,
-    transactions::utils::update_account, WebStore,
+    chain_data::utils::serialize_chain_mmr_node,
+    notes::utils::{upsert_input_note_tx, upsert_output_note_tx},
+    transactions::utils::update_account,
+    WebStore,
 };
 use crate::{
-    store::{CommittedNoteState, InputNoteRecord, NoteFilter, StoreError},
+    store::{
+        input_note_states::CommittedNoteState, InputNoteRecord, NoteFilter, OutputNoteRecord,
+        StoreError,
+    },
     sync::{NoteTagRecord, NoteTagSource, StateSyncUpdate},
 };
 
@@ -25,9 +30,6 @@ use js_bindings::*;
 
 mod models;
 use models::*;
-
-mod flattened_vec;
-use flattened_vec::*;
 
 impl WebStore {
     pub(crate) async fn get_note_tags(&self) -> Result<Vec<NoteTagRecord>, StoreError> {
@@ -102,8 +104,8 @@ impl WebStore {
         &self,
         state_sync_update: StateSyncUpdate,
     ) -> Result<(), StoreError> {
-        let mut relevant_input_notes =
-            self.get_relevant_sync_input_notes(&state_sync_update).await.unwrap();
+        let (mut relevant_input_notes, mut relevant_output_notes) =
+            self.get_relevant_sync_notes(&state_sync_update).await?;
 
         let StateSyncUpdate {
             block_header,
@@ -119,16 +121,6 @@ impl WebStore {
         // Serialize data for updating state sync and block header
         let block_num_as_str = block_header.block_num().to_string();
 
-        // Serialize data for updating spent notes
-        let nullifiers_as_str = nullifiers
-            .iter()
-            .map(|nullifier_update| nullifier_update.nullifier.to_hex())
-            .collect();
-        let nullifier_block_nums_as_str = nullifiers
-            .iter()
-            .map(|nullifier_update| nullifier_update.block_num.to_string())
-            .collect();
-
         // Serialize data for updating block header
         let block_header_as_bytes = block_header.to_bytes();
         let new_mmr_peaks_as_bytes = new_mmr_peaks.peaks().to_vec().to_bytes();
@@ -142,28 +134,15 @@ impl WebStore {
             serialized_nodes.push(serialized_data.node);
         }
 
-        // Serialize data for updating committed notes
-        let output_note_ids_as_str: Vec<String> = committed_notes
-            .updated_output_notes()
-            .iter()
-            .map(|(note_id, _)| note_id.inner().to_hex())
-            .collect();
-
-        let output_note_inclusion_proofs_as_bytes: Vec<Vec<u8>> = committed_notes
-            .updated_output_notes()
-            .iter()
-            .map(|(_, inclusion_proof)| {
-                let block_num = inclusion_proof.location().block_num();
-                let note_index = inclusion_proof.location().node_index_in_block();
-
-                // Create a NoteInclusionProof and serialize it to JSON, handle errors with `?`
-                NoteInclusionProof::new(block_num, note_index, inclusion_proof.note_path().clone())
-                    .unwrap()
-                    .to_bytes()
-            })
-            .collect::<Vec<Vec<u8>>>();
-        let flattened_nested_vec_output_note_inclusion_proofs =
-            flatten_nested_u8_vec(output_note_inclusion_proofs_as_bytes);
+        for (note_id, inclusion_proof) in committed_notes.updated_output_notes().iter() {
+            if let Some(output_note_record) =
+                relevant_output_notes.iter_mut().find(|n| n.id() == *note_id)
+            {
+                if output_note_record.inclusion_proof_received(inclusion_proof.clone())? {
+                    upsert_output_note_tx(output_note_record).await?;
+                }
+            }
+        }
 
         let input_note_ids_as_str: Vec<String> = committed_notes
             .updated_input_notes()
@@ -187,7 +166,7 @@ impl WebStore {
                     input_note_record.block_header_received(block_header)?;
 
                 if inclusion_proof_received || block_header_received {
-                    upsert_input_note_tx(input_note_record.clone()).await.unwrap();
+                    upsert_input_note_tx(input_note_record).await.unwrap();
                 }
             }
         }
@@ -211,7 +190,45 @@ impl WebStore {
                 .into(),
             );
 
-            upsert_input_note_tx(input_note_record).await.unwrap();
+            upsert_input_note_tx(&input_note_record).await.unwrap();
+        }
+
+        // Committed transactions
+        for transaction_update in committed_transactions.iter() {
+            if let Some(input_note_record) = relevant_input_notes.iter_mut().find(|n| {
+                n.is_processing()
+                    && n.consumer_transaction_id() == Some(&transaction_update.transaction_id)
+            }) {
+                if input_note_record.transaction_committed(
+                    transaction_update.transaction_id,
+                    transaction_update.block_num,
+                )? {
+                    upsert_input_note_tx(input_note_record).await?;
+                }
+            }
+        }
+
+        // Update spent notes
+        for nullifier_update in nullifiers.iter() {
+            let nullifier = nullifier_update.nullifier;
+            let block_num = nullifier_update.block_num;
+
+            if let Some(input_note_record) =
+                relevant_input_notes.iter_mut().find(|n| n.nullifier() == nullifier)
+            {
+                if input_note_record.consumed_externally(nullifier, block_num)? {
+                    upsert_input_note_tx(input_note_record).await?;
+                }
+            }
+
+            if let Some(output_note_record) = relevant_output_notes
+                .iter_mut()
+                .find(|n| n.nullifier().is_some_and(|n| n == nullifier))
+            {
+                if output_note_record.nullifier_received(nullifier, block_num)? {
+                    upsert_output_note_tx(output_note_record).await?;
+                }
+            }
         }
 
         // Serialize data for updating committed transactions
@@ -232,15 +249,11 @@ impl WebStore {
 
         let promise = idxdb_apply_state_sync(
             block_num_as_str,
-            nullifiers_as_str,
-            nullifier_block_nums_as_str,
             block_header_as_bytes,
             new_mmr_peaks_as_bytes,
             block_has_relevant_notes,
             serialized_node_ids,
             serialized_nodes,
-            output_note_ids_as_str,
-            flattened_nested_vec_output_note_inclusion_proofs,
             input_note_ids_as_str,
             transactions_to_commit_as_str,
             transactions_to_commit_block_nums_as_str,
@@ -250,12 +263,12 @@ impl WebStore {
         Ok(())
     }
 
-    /// Get the input notes from the store that are relevant to the state sync update. Secifically,
+    /// Get the notes from the store that are relevant to the state sync update. Secifically,
     /// notes that were updated and nullified during the sync.
-    async fn get_relevant_sync_input_notes(
+    async fn get_relevant_sync_notes(
         &self,
         state_sync_update: &StateSyncUpdate,
-    ) -> Result<Vec<InputNoteRecord>, StoreError> {
+    ) -> Result<(Vec<InputNoteRecord>, Vec<OutputNoteRecord>), StoreError> {
         let StateSyncUpdate { nullifiers, synced_new_notes, .. } = state_sync_update;
 
         let updated_input_note_ids = synced_new_notes
@@ -263,19 +276,34 @@ impl WebStore {
             .iter()
             .map(|input_note| input_note.id())
             .collect::<Vec<_>>();
-        let updated_input_notes =
-            self.get_input_notes(NoteFilter::List(updated_input_note_ids)).await.unwrap();
 
-        let nullifiers = nullifiers
+        let updated_output_note_ids = synced_new_notes
+            .updated_output_notes()
+            .iter()
+            .map(|output_note| output_note.0)
+            .collect::<Vec<_>>();
+
+        let updated_input_notes =
+            self.get_input_notes(NoteFilter::List(updated_input_note_ids)).await?;
+        let updated_output_notes =
+            self.get_output_notes(NoteFilter::List(updated_output_note_ids)).await?;
+
+        let nullifiers: Vec<Nullifier> = nullifiers
             .iter()
             .map(|nullifier_update| nullifier_update.nullifier)
             .collect::<Vec<_>>();
-        let nullified_notes =
-            self.get_input_notes(NoteFilter::Nullifiers(nullifiers)).await.unwrap();
 
-        let mut relevant_notes = updated_input_notes;
-        relevant_notes.extend(nullified_notes);
+        let nullified_input_notes =
+            self.get_input_notes(NoteFilter::Nullifiers(nullifiers.clone())).await?;
+        let nullified_output_notes =
+            self.get_output_notes(NoteFilter::Nullifiers(nullifiers)).await?;
 
-        Ok(relevant_notes)
+        let mut relevant_input_notes = updated_input_notes;
+        relevant_input_notes.extend(nullified_input_notes);
+
+        let mut relevant_output_notes = updated_output_notes;
+        relevant_output_notes.extend(nullified_output_notes);
+
+        Ok((relevant_input_notes, relevant_output_notes))
     }
 }